test(embed): relocate pooling agreement test out of lattice-inference
Move the pooling-agreement test into lattice-embed, where EmbeddingModel
and lattice_inference::pool are the same compiled unit, so the test can
compare BertPooling values directly instead of their Debug strings.
Removes the dev-only lattice-embed dependency this test previously
required inside lattice-inference.
