+++ crates/embed/src/model.rs
@@ -755,4 +755,102 @@ mod tests {
             "BGE and E5 must use different pooling strategies"
         );
     }
+
+    /// Guards `lattice_inference::pool::default_bert_pooling_for_model_name`
+    /// (the served-directory-name lookup `lattice_serve` uses) against
+    /// drifting from `EmbeddingModel::bert_pooling` (this table). The
+    /// expectation comes from calling `bert_pooling()` itself, never from
+    /// retyping its rule, so a wrong mirror cannot agree with a wrong
+    /// expectation.
+    ///
+    /// This lives in `lattice-embed`, not `lattice-inference`, so both sides
+    /// of the comparison are the same compiled unit: `lattice-embed` already
+    /// depends on `lattice-inference` (this crate's own `[dependencies]`,
+    /// gated by the `native` feature), so `EmbeddingModel::bert_pooling()`'s
+    /// return type and `default_bert_pooling_for_model_name`'s return type
+    /// are literally the same `lattice_inference::BertPooling`, comparable
+    /// with `==` directly -- no `format!("{:?}")` string workaround needed.
+    ///
+    /// Enumeration: `EmbeddingModel` has no `EnumIter`/`VariantArray`/
+    /// `ALL_MODELS` constant (checked via `grep -n
+    /// "strum\|EnumIter\|VariantArray\|ALL_MODELS" crates/embed/src/model.rs`,
+    /// no hits), so `variants` below is still a hand-transcribed list of all
+    /// 10 current variants. But `#[non_exhaustive]` only forces a wildcard
+    /// arm in matches written *outside* the declaring crate -- this test
+    /// lives inside `lattice-embed` itself, where `EmbeddingModel` is
+    /// declared, so a wildcard-free match over it type-checks here. The
+    /// `assert_variants_exhaustively_covered` match below has no wildcard
+    /// arm; it is checked against the enum's full current variant set
+    /// regardless of which values are actually passed to it, so a variant
+    /// added to `EmbeddingModel` without a matching arm here is a compile
+    /// error ("non-exhaustive patterns"), not a silent gap in `variants`.
+    #[cfg(feature = "native")]
+    #[test]
+    fn default_bert_pooling_for_model_name_matches_embedding_model_bert_pooling_table() {
+        use lattice_inference::pool::default_bert_pooling_for_model_name;
+
+        fn assert_variants_exhaustively_covered(m: &EmbeddingModel) {
+            match m {
+                EmbeddingModel::BgeSmallEnV15
+                | EmbeddingModel::BgeBaseEnV15
+                | EmbeddingModel::BgeLargeEnV15
+                | EmbeddingModel::MultilingualE5Small
+                | EmbeddingModel::MultilingualE5Base
+                | EmbeddingModel::AllMiniLmL6V2
+                | EmbeddingModel::ParaphraseMultilingualMiniLmL12V2
+                | EmbeddingModel::Qwen3Embedding0_6B
+                | EmbeddingModel::Qwen3Embedding4B
+                | EmbeddingModel::TextEmbedding3Small => {}
+            }
+        }
+
+        let variants = [
+            EmbeddingModel::BgeSmallEnV15,
+            EmbeddingModel::BgeBaseEnV15,
+            EmbeddingModel::BgeLargeEnV15,
+            EmbeddingModel::MultilingualE5Small,
+            EmbeddingModel::MultilingualE5Base,
+            EmbeddingModel::AllMiniLmL6V2,
+            EmbeddingModel::ParaphraseMultilingualMiniLmL12V2,
+            EmbeddingModel::Qwen3Embedding0_6B,
+            EmbeddingModel::Qwen3Embedding4B,
+            EmbeddingModel::TextEmbedding3Small,
+        ];
+
+        let mut bert_family_checked = 0usize;
+        for variant in variants {
+            assert_variants_exhaustively_covered(&variant);
+
+            let Some(expected) = variant.bert_pooling() else {
+                continue;
+            };
+            bert_family_checked += 1;
+            // Same shape of value production reaches
+            // `default_bert_pooling_for_model_name` in production:
+            // `lattice_serve` passes it the served directory's last path
+            // component (`resolve_model_dir(...).file_name()`), not the
+            // model's full HF id. Simulate that here by taking the last
+            // path component of the model's own `model_id()`.
+            let dir_name = std::path::Path::new(variant.model_id())
+                .file_name()
+                .and_then(|n| n.to_str())
+                .expect("model_id() yields a valid path component");
+            let actual = default_bert_pooling_for_model_name(dir_name);
+            assert_eq!(
+                actual, expected,
+                "pooling mismatch for {variant:?} (derived dir name {dir_name:?})"
+            );
+        }
+        // BGE small/base/large + E5 small/base + MiniLM + paraphrase-MiniLM:
+        // the 7 variants whose `bert_pooling()` returns `Some(_)` as of this
+        // writing. Qwen3 x2 and TextEmbedding3Small return `None` and are
+        // skipped above. This catches a variant's `bert_pooling()` gate
+        // moving, but -- per the enumeration note above -- a wholly new
+        // variant absent from `variants` only surfaces as the compile error
+        // from `assert_variants_exhaustively_covered`, not from this count.
+        assert_eq!(
+            bert_family_checked, 7,
+            "EmbeddingModel's BERT-family (Some(_) from bert_pooling()) variant count changed"
+        );
+    }
 }
